Logs over RPC w/ lots to touch up
