+++ client/allocdir/alloc_dir.go
@@ -2,6 +2,7 @@ package allocdir
 
 import (
 	"archive/tar"
+	"context"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -10,11 +11,10 @@ import (
 	"path/filepath"
 	"time"
 
-	"gopkg.in/tomb.v1"
-
 	"github.com/hashicorp/go-multierror"
 	"github.com/hashicorp/nomad/nomad/structs"
 	"github.com/hpcloud/tail/watch"
+	tomb "gopkg.in/tomb.v1"
 )
 
 const (
@@ -90,8 +90,8 @@ type AllocDirFS interface {
 	Stat(path string) (*AllocFileInfo, error)
 	ReadAt(path string, offset int64) (io.ReadCloser, error)
 	Snapshot(w io.Writer) error
-	BlockUntilExists(path string, t *tomb.Tomb) (chan error, error)
-	ChangeEvents(path string, curOffset int64, t *tomb.Tomb) (*watch.FileChanges, error)
+	BlockUntilExists(ctx context.Context, path string) (chan error, error)
+	ChangeEvents(ctx context.Context, path string, curOffset int64) (*watch.FileChanges, error)
 }
 
 // NewAllocDir initializes the AllocDir struct with allocDir as base path for
@@ -411,8 +411,8 @@ func (d *AllocDir) ReadAt(path string, offset int64) (io.ReadCloser, error) {
 }
 
 // BlockUntilExists blocks until the passed file relative the allocation
-// directory exists. The block can be cancelled with the passed tomb.
-func (d *AllocDir) BlockUntilExists(path string, t *tomb.Tomb) (chan error, error) {
+// directory exists. The block can be cancelled with the passed context.
+func (d *AllocDir) BlockUntilExists(ctx context.Context, path string) (chan error, error) {
 	if escapes, err := structs.PathEscapesAllocDir("", path); err != nil {
 		return nil, fmt.Errorf("Failed to check if path escapes alloc directory: %v", err)
 	} else if escapes {
@@ -423,6 +423,11 @@ func (d *AllocDir) BlockUntilExists(path string, t *tomb.Tomb) (chan error, erro
 	p := filepath.Join(d.AllocDir, path)
 	watcher := getFileWatcher(p)
 	returnCh := make(chan error, 1)
+	t := &tomb.Tomb{}
+	go func() {
+		<-ctx.Done()
+		t.Kill(nil)
+	}()
 	go func() {
 		returnCh <- watcher.BlockUntilExists(t)
 		close(returnCh)
@@ -431,14 +436,20 @@ func (d *AllocDir) BlockUntilExists(path string, t *tomb.Tomb) (chan error, erro
 }
 
 // ChangeEvents watches for changes to the passed path relative to the
-// allocation directory. The offset should be the last read offset. The tomb is
+// allocation directory. The offset should be the last read offset. The context is
 // used to clean up the watch.
-func (d *AllocDir) ChangeEvents(path string, curOffset int64, t *tomb.Tomb) (*watch.FileChanges, error) {
+func (d *AllocDir) ChangeEvents(ctx context.Context, path string, curOffset int64) (*watch.FileChanges, error) {
 	if escapes, err := structs.PathEscapesAllocDir("", path); err != nil {
 		return nil, fmt.Errorf("Failed to check if path escapes alloc directory: %v", err)
 	} else if escapes {
 		return nil, fmt.Errorf("Path escapes the alloc directory")
 	}
+
+	t := &tomb.Tomb{}
+	go func() {
+		<-ctx.Done()
+		t.Kill(nil)
+	}()
 
 	// Get the path relative to the alloc directory
 	p := filepath.Join(d.AllocDir, path)
