+++ command/agent/fs_endpoint_test.go
@@ -1,31 +1,27 @@
 package agent
 
 import (
-	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"log"
-	"math"
 	"net/http"
 	"net/http/httptest"
-	"os"
-	"path/filepath"
-	"reflect"
-	"runtime"
-	"strconv"
-	"strings"
 	"testing"
-	"time"
 
 	"github.com/hashicorp/nomad/acl"
-	"github.com/hashicorp/nomad/client/allocdir"
 	"github.com/hashicorp/nomad/nomad/mock"
 	"github.com/hashicorp/nomad/nomad/structs"
-	"github.com/hashicorp/nomad/testutil"
 	"github.com/stretchr/testify/assert"
-	"github.com/ugorji/go/codec"
 )
+
+type WriteCloseChecker struct {
+	io.WriteCloser
+	Closed bool
+}
+
+func (w *WriteCloseChecker) Close() error {
+	w.Closed = true
+	return w.WriteCloser.Close()
+}
 
 func TestAllocDirFS_List_MissingParams(t *testing.T) {
 	t.Parallel()
@@ -229,402 +225,7 @@ func TestAllocDirFS_Logs_ACL(t *testing.T) {
 	})
 }
 
-type WriteCloseChecker struct {
-	io.WriteCloser
-	Closed bool
-}
-
-func (w *WriteCloseChecker) Close() error {
-	w.Closed = true
-	return w.WriteCloser.Close()
-}
-
-// This test checks, that even if the frame size has not been hit, a flush will
-// periodically occur.
-func TestStreamFramer_Flush(t *testing.T) {
-	// Create the stream framer
-	r, w := io.Pipe()
-	wrappedW := &WriteCloseChecker{WriteCloser: w}
-	hRate, bWindow := 100*time.Millisecond, 100*time.Millisecond
-	sf := NewStreamFramer(wrappedW, false, hRate, bWindow, 100)
-	sf.Run()
-
-	// Create a decoder
-	dec := codec.NewDecoder(r, structs.JsonHandle)
-
-	f := "foo"
-	fe := "bar"
-	d := []byte{0xa}
-	o := int64(10)
-
-	// Start the reader
-	resultCh := make(chan struct{})
-	go func() {
-		for {
-			var frame StreamFrame
-			if err := dec.Decode(&frame); err != nil {
-				t.Fatalf("failed to decode")
-			}
-
-			if frame.IsHeartbeat() {
-				continue
-			}
-
-			if reflect.DeepEqual(frame.Data, d) && frame.Offset == o && frame.File == f && frame.FileEvent == fe {
-				resultCh <- struct{}{}
-				return
-			}
-
-		}
-	}()
-
-	// Write only 1 byte so we do not hit the frame size
-	if err := sf.Send(f, fe, d, o); err != nil {
-		t.Fatalf("Send() failed %v", err)
-	}
-
-	select {
-	case <-resultCh:
-	case <-time.After(10 * time.Duration(testutil.TestMultiplier()) * bWindow):
-		t.Fatalf("failed to flush")
-	}
-
-	// Close the reader and wait. This should cause the runner to exit
-	if err := r.Close(); err != nil {
-		t.Fatalf("failed to close reader")
-	}
-
-	select {
-	case <-sf.ExitCh():
-	case <-time.After(10 * time.Duration(testutil.TestMultiplier()) * hRate):
-		t.Fatalf("exit channel should close")
-	}
-
-	sf.Destroy()
-	if !wrappedW.Closed {
-		t.Fatalf("writer not closed")
-	}
-}
-
-// This test checks that frames will be batched till the frame size is hit (in
-// the case that is before the flush).
-func TestStreamFramer_Batch(t *testing.T) {
-	// Create the stream framer
-	r, w := io.Pipe()
-	wrappedW := &WriteCloseChecker{WriteCloser: w}
-	// Ensure the batch window doesn't get hit
-	hRate, bWindow := 100*time.Millisecond, 500*time.Millisecond
-	sf := NewStreamFramer(wrappedW, false, hRate, bWindow, 3)
-	sf.Run()
-
-	// Create a decoder
-	dec := codec.NewDecoder(r, structs.JsonHandle)
-
-	f := "foo"
-	fe := "bar"
-	d := []byte{0xa, 0xb, 0xc}
-	o := int64(10)
-
-	// Start the reader
-	resultCh := make(chan struct{})
-	go func() {
-		for {
-			var frame StreamFrame
-			if err := dec.Decode(&frame); err != nil {
-				t.Fatalf("failed to decode")
-			}
-
-			if frame.IsHeartbeat() {
-				continue
-			}
-
-			if reflect.DeepEqual(frame.Data, d) && frame.Offset == o && frame.File == f && frame.FileEvent == fe {
-				resultCh <- struct{}{}
-				return
-			}
-		}
-	}()
-
-	// Write only 1 byte so we do not hit the frame size
-	if err := sf.Send(f, fe, d[:1], o); err != nil {
-		t.Fatalf("Send() failed %v", err)
-	}
-
-	// Ensure we didn't get any data
-	select {
-	case <-resultCh:
-		t.Fatalf("Got data before frame size reached")
-	case <-time.After(bWindow / 2):
-	}
-
-	// Write the rest so we hit the frame size
-	if err := sf.Send(f, fe, d[1:], o); err != nil {
-		t.Fatalf("Send() failed %v", err)
-	}
-
-	// Ensure we get data
-	select {
-	case <-resultCh:
-	case <-time.After(10 * time.Duration(testutil.TestMultiplier()) * bWindow):
-		t.Fatalf("Did not receive data after batch size reached")
-	}
-
-	// Close the reader and wait. This should cause the runner to exit
-	if err := r.Close(); err != nil {
-		t.Fatalf("failed to close reader")
-	}
-
-	select {
-	case <-sf.ExitCh():
-	case <-time.After(10 * time.Duration(testutil.TestMultiplier()) * hRate):
-		t.Fatalf("exit channel should close")
-	}
-
-	sf.Destroy()
-	if !wrappedW.Closed {
-		t.Fatalf("writer not closed")
-	}
-}
-
-func TestStreamFramer_Heartbeat(t *testing.T) {
-	// Create the stream framer
-	r, w := io.Pipe()
-	wrappedW := &WriteCloseChecker{WriteCloser: w}
-	hRate, bWindow := 100*time.Millisecond, 100*time.Millisecond
-	sf := NewStreamFramer(wrappedW, false, hRate, bWindow, 100)
-	sf.Run()
-
-	// Create a decoder
-	dec := codec.NewDecoder(r, structs.JsonHandle)
-
-	// Start the reader
-	resultCh := make(chan struct{})
-	go func() {
-		for {
-			var frame StreamFrame
-			if err := dec.Decode(&frame); err != nil {
-				t.Fatalf("failed to decode")
-			}
-
-			if frame.IsHeartbeat() {
-				resultCh <- struct{}{}
-				return
-			}
-		}
-	}()
-
-	select {
-	case <-resultCh:
-	case <-time.After(10 * time.Duration(testutil.TestMultiplier()) * hRate):
-		t.Fatalf("failed to heartbeat")
-	}
-
-	// Close the reader and wait. This should cause the runner to exit
-	if err := r.Close(); err != nil {
-		t.Fatalf("failed to close reader")
-	}
-
-	select {
-	case <-sf.ExitCh():
-	case <-time.After(10 * time.Duration(testutil.TestMultiplier()) * hRate):
-		t.Fatalf("exit channel should close")
-	}
-
-	sf.Destroy()
-	if !wrappedW.Closed {
-		t.Fatalf("writer not closed")
-	}
-}
-
-// This test checks that frames are received in order
-func TestStreamFramer_Order(t *testing.T) {
-	// Create the stream framer
-	r, w := io.Pipe()
-	wrappedW := &WriteCloseChecker{WriteCloser: w}
-	// Ensure the batch window doesn't get hit
-	hRate, bWindow := 100*time.Millisecond, 10*time.Millisecond
-	sf := NewStreamFramer(wrappedW, false, hRate, bWindow, 10)
-	sf.Run()
-
-	// Create a decoder
-	dec := codec.NewDecoder(r, structs.JsonHandle)
-
-	files := []string{"1", "2", "3", "4", "5"}
-	input := bytes.NewBuffer(make([]byte, 0, 100000))
-	for i := 0; i <= 1000; i++ {
-		str := strconv.Itoa(i) + ","
-		input.WriteString(str)
-	}
-
-	expected := bytes.NewBuffer(make([]byte, 0, 100000))
-	for range files {
-		expected.Write(input.Bytes())
-	}
-	receivedBuf := bytes.NewBuffer(make([]byte, 0, 100000))
-
-	// Start the reader
-	resultCh := make(chan struct{})
-	go func() {
-		for {
-			var frame StreamFrame
-			if err := dec.Decode(&frame); err != nil {
-				t.Fatalf("failed to decode")
-			}
-
-			if frame.IsHeartbeat() {
-				continue
-			}
-
-			receivedBuf.Write(frame.Data)
-
-			if reflect.DeepEqual(expected, receivedBuf) {
-				resultCh <- struct{}{}
-				return
-			}
-		}
-	}()
-
-	// Send the data
-	b := input.Bytes()
-	shards := 10
-	each := len(b) / shards
-	for _, f := range files {
-		for i := 0; i < shards; i++ {
-			l, r := each*i, each*(i+1)
-			if i == shards-1 {
-				r = len(b)
-			}
-
-			if err := sf.Send(f, "", b[l:r], 0); err != nil {
-				t.Fatalf("Send() failed %v", err)
-			}
-		}
-	}
-
-	// Ensure we get data
-	select {
-	case <-resultCh:
-	case <-time.After(10 * time.Duration(testutil.TestMultiplier()) * bWindow):
-		if reflect.DeepEqual(expected, receivedBuf) {
-			got := receivedBuf.String()
-			want := expected.String()
-			t.Fatalf("Got %v; want %v", got, want)
-		}
-	}
-
-	// Close the reader and wait. This should cause the runner to exit
-	if err := r.Close(); err != nil {
-		t.Fatalf("failed to close reader")
-	}
-
-	select {
-	case <-sf.ExitCh():
-	case <-time.After(10 * time.Duration(testutil.TestMultiplier()) * hRate):
-		t.Fatalf("exit channel should close")
-	}
-
-	sf.Destroy()
-	if !wrappedW.Closed {
-		t.Fatalf("writer not closed")
-	}
-}
-
-// This test checks that frames are received in order
-func TestStreamFramer_Order_PlainText(t *testing.T) {
-	// Create the stream framer
-	r, w := io.Pipe()
-	wrappedW := &WriteCloseChecker{WriteCloser: w}
-	// Ensure the batch window doesn't get hit
-	hRate, bWindow := 100*time.Millisecond, 10*time.Millisecond
-	sf := NewStreamFramer(wrappedW, true, hRate, bWindow, 10)
-	sf.Run()
-
-	files := []string{"1", "2", "3", "4", "5"}
-	input := bytes.NewBuffer(make([]byte, 0, 100000))
-	for i := 0; i <= 1000; i++ {
-		str := strconv.Itoa(i) + ","
-		input.WriteString(str)
-	}
-
-	expected := bytes.NewBuffer(make([]byte, 0, 100000))
-	for range files {
-		expected.Write(input.Bytes())
-	}
-	receivedBuf := bytes.NewBuffer(make([]byte, 0, 100000))
-
-	// Start the reader
-	resultCh := make(chan struct{})
-	go func() {
-	OUTER:
-		for {
-			if _, err := receivedBuf.ReadFrom(r); err != nil {
-				if strings.Contains(err.Error(), "closed pipe") {
-					resultCh <- struct{}{}
-					return
-				}
-				t.Fatalf("bad read: %v", err)
-			}
-
-			if expected.Len() != receivedBuf.Len() {
-				continue
-			}
-			expectedBytes := expected.Bytes()
-			actualBytes := receivedBuf.Bytes()
-			for i, e := range expectedBytes {
-				if a := actualBytes[i]; a != e {
-					continue OUTER
-				}
-			}
-			resultCh <- struct{}{}
-			return
-
-		}
-	}()
-
-	// Send the data
-	b := input.Bytes()
-	shards := 10
-	each := len(b) / shards
-	for _, f := range files {
-		for i := 0; i < shards; i++ {
-			l, r := each*i, each*(i+1)
-			if i == shards-1 {
-				r = len(b)
-			}
-
-			if err := sf.Send(f, "", b[l:r], 0); err != nil {
-				t.Fatalf("Send() failed %v", err)
-			}
-		}
-	}
-
-	// Ensure we get data
-	select {
-	case <-resultCh:
-	case <-time.After(10 * time.Duration(testutil.TestMultiplier()) * bWindow):
-		if expected.Len() != receivedBuf.Len() {
-			t.Fatalf("Got %v; want %v", expected.Len(), receivedBuf.Len())
-		}
-		expectedBytes := expected.Bytes()
-		actualBytes := receivedBuf.Bytes()
-		for i, e := range expectedBytes {
-			if a := actualBytes[i]; a != e {
-				t.Fatalf("Index %d; Got %q; want %q", i, a, e)
-			}
-		}
-	}
-
-	// Close the reader and wait. This should cause the runner to exit
-	if err := r.Close(); err != nil {
-		t.Fatalf("failed to close reader")
-	}
-
-	sf.Destroy()
-	if !wrappedW.Closed {
-		t.Fatalf("writer not closed")
-	}
-}
-
+/*
 func TestHTTP_Stream_MissingParams(t *testing.T) {
 	t.Parallel()
 	httpTest(t, nil, func(s *TestAgent) {
@@ -693,7 +294,7 @@ func TestHTTP_Stream_NoFile(t *testing.T) {
 		ad := tempAllocDir(t)
 		defer os.RemoveAll(ad.AllocDir)
 
-		framer := NewStreamFramer(nopWriteCloser{ioutil.Discard}, false, streamHeartbeatRate, streamBatchWindow, streamFrameSize)
+		framer := sframer.NewStreamFramer(nopWriteCloser{ioutil.Discard}, false, streamHeartbeatRate, streamBatchWindow, streamFrameSize)
 		framer.Run()
 		defer framer.Destroy()
 
@@ -731,7 +332,7 @@ func TestHTTP_Stream_Modify(t *testing.T) {
 		go func() {
 			var collected []byte
 			for {
-				var frame StreamFrame
+				var frame sframer.StreamFrame
 				if err := dec.Decode(&frame); err != nil {
 					t.Fatalf("failed to decode: %v", err)
 				}
@@ -753,7 +354,7 @@ func TestHTTP_Stream_Modify(t *testing.T) {
 			t.Fatalf("write failed: %v", err)
 		}
 
-		framer := NewStreamFramer(w, false, streamHeartbeatRate, streamBatchWindow, streamFrameSize)
+		framer := sframer.NewStreamFramer(w, false, streamHeartbeatRate, streamBatchWindow, streamFrameSize)
 		framer.Run()
 		defer framer.Destroy()
 
@@ -809,7 +410,7 @@ func TestHTTP_Stream_Truncate(t *testing.T) {
 		go func() {
 			var collected []byte
 			for {
-				var frame StreamFrame
+				var frame sframer.StreamFrame
 				if err := dec.Decode(&frame); err != nil {
 					t.Fatalf("failed to decode: %v", err)
 				}
@@ -835,7 +436,7 @@ func TestHTTP_Stream_Truncate(t *testing.T) {
 			t.Fatalf("write failed: %v", err)
 		}
 
-		framer := NewStreamFramer(w, false, streamHeartbeatRate, streamBatchWindow, streamFrameSize)
+		framer := sframer.NewStreamFramer(w, false, streamHeartbeatRate, streamBatchWindow, streamFrameSize)
 		framer.Run()
 		defer framer.Destroy()
 
@@ -918,7 +519,7 @@ func TestHTTP_Stream_Delete(t *testing.T) {
 		deleteCh := make(chan struct{})
 		go func() {
 			for {
-				var frame StreamFrame
+				var frame sframer.StreamFrame
 				if err := dec.Decode(&frame); err != nil {
 					t.Fatalf("failed to decode: %v", err)
 				}
@@ -939,7 +540,7 @@ func TestHTTP_Stream_Delete(t *testing.T) {
 			t.Fatalf("write failed: %v", err)
 		}
 
-		framer := NewStreamFramer(wrappedW, false, streamHeartbeatRate, streamBatchWindow, streamFrameSize)
+		framer := sframer.NewStreamFramer(wrappedW, false, streamHeartbeatRate, streamBatchWindow, streamFrameSize)
 		framer.Run()
 
 		// Start streaming
@@ -1010,7 +611,7 @@ func TestHTTP_Logs_NoFollow(t *testing.T) {
 		resultCh := make(chan struct{})
 		go func() {
 			for {
-				var frame StreamFrame
+				var frame sframer.StreamFrame
 				if err := dec.Decode(&frame); err != nil {
 					if err == io.EOF {
 						t.Logf("EOF")
@@ -1098,7 +699,7 @@ func TestHTTP_Logs_Follow(t *testing.T) {
 		fullResultCh := make(chan struct{})
 		go func() {
 			for {
-				var frame StreamFrame
+				var frame sframer.StreamFrame
 				if err := dec.Decode(&frame); err != nil {
 					if err == io.EOF {
 						t.Logf("EOF")
@@ -1213,7 +814,7 @@ func BenchmarkHTTP_Logs_Follow(t *testing.B) {
 		fullResultCh := make(chan struct{})
 		go func() {
 			for {
-				var frame StreamFrame
+				var frame sframer.StreamFrame
 				if err := dec.Decode(&frame); err != nil {
 					if err == io.EOF {
 						t.Logf("EOF")
@@ -1468,3 +1069,4 @@ func TestLogs_findClosest(t *testing.T) {
 		}
 	}
 }
+*/
